Move view/background to view/oldbackground, BackgroundConfig to model, and new backgrounds classes to view/background
