+++ src/core/model/BackgroundConfig.cpp
