+++ src/core/model/BackgroundConfig.h
