+++ src/core/view/background/ImageBackgroundView.cpp
