+++ src/core/view/background/ImageBackgroundView.h
