+++ src/core/view/oldbackground/BaseBackgroundPainter.cpp
