+++ src/core/view/oldbackground/DottedBackgroundPainter.cpp
