+++ src/core/view/oldbackground/DottedBackgroundPainter.h
