+++ src/core/view/oldbackground/GraphBackgroundPainter.cpp
