+++ src/core/view/oldbackground/GraphBackgroundPainter.h
