+++ src/core/view/oldbackground/IsometricBackgroundPainter.cpp
