+++ src/core/view/oldbackground/IsometricBackgroundPainter.h
