+++ src/core/view/oldbackground/LineBackgroundPainter.cpp
