+++ src/core/view/oldbackground/LineBackgroundPainter.h
