+++ src/core/view/oldbackground/MainBackgroundPainter.h
