+++ src/core/view/oldbackground/StavesBackgroundPainter.cpp
