+++ src/core/view/oldbackground/StavesBackgroundPainter.h
